+++ js/constants.js
@@ -1,0 +1,58 @@
+export const DESCRIPTIONS = [
+  'Утро',
+  'Солнышко',
+  'Котенок',
+];
+
+export const PHOTOS_COUNT = 25;
+
+export const MIN_AVATAR = 1;
+export const MAX_AVATAR = 6;
+
+export const MIN_ID = 0;
+export const MAX_ID = 1000000;
+
+export const MIN_LIKES = 15;
+export const MAX_LIKES = 200;
+
+export const MIN_COMMENTS = 0;
+export const MAX_COMMENTS = 30;
+
+
+export const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра.' ,
+  'В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают.' ,
+  'Как можно было поймать такой неудачный момент?!',
+];
+
+export const NAMES = [
+  'Алексей',
+  'Иван',
+  'Дмитрий',
+  'Юлия',
+  'Андрей',
+  'Николай',
+  'Михаил',
+  'Егор',
+  'Артем',
+  'Владимир',
+  'Александр',
+  'Максим',
+  'Олег',
+  'Юрий',
+  'Павел',
+  'Саша',
+  'Светлана',
+  'Вика',
+  'Маша',
+  'Елена',
+  'Оля',
+  'Татьяна',
+  'Наталья',
+  'Анна',
+];
